+++ R/tables.R
@@ -12,27 +12,27 @@
 #' @param stages A list of stages for columns of the table. Default is `ReboundTools::rebound_stages`.
 #'               Stage order is preserved in the table.
 #' @param latex_stages See `ReboundTools::latex_rebound_stages`. Set `NULL` to prevent conversion to LaTeX stage names.
-#' @param file An optional path to a file. Default is `sample_eeu_data_path()`.
 #' @param case See `ReboundTools::eeu_base_params`.
 #' @param service_unit,energy_engr_unit See `ReboundTools::eeu_base_params`.
 #' @param ... Arguments passed to `xtable::xtable()`, possibly
 #'            `label`, `caption`, `digits`, etc.
 #' @param .var,.stage,.var_stage,.value,.name,.unit_col Column names used internally.
 #'
-#' @return An `xtable` object giving the details of the table.
+#' @return An `xtable` object suitable for printing.
 #' 
 #' @export
 #'
 #' @examples
-#' stages_table()
-stages_table <- function(.analysis_data = rebound_analysis(load_eeu_data(file)), 
+#' load_eeu_data() %>% 
+#'   rebound_analysis() %>% 
+#'   stages_table()
+stages_table <- function(.analysis_data, 
                          add_units = TRUE,
                          escape_latex = TRUE,
                          vars = ReboundTools::key_analysis_vars, 
                          latex_vars = ReboundTools::latex_key_analysis_vars,
                          stages = ReboundTools::rebound_stages, 
                          latex_stages = ReboundTools::latex_rebound_stages,
-                         file = sample_eeu_data_path(), 
                          case = ReboundTools::eeu_base_params$case, 
                          service_unit = ReboundTools::eeu_base_params$service_unit,
                          energy_engr_unit = ReboundTools::eeu_base_params$energy_engr_unit,
@@ -145,16 +145,114 @@ stages_table <- function(.analysis_data = rebound_analysis(load_eeu_data(file)),
 }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
+#' Build a rebound results table 
+#'
+#' @param .analysis_data Rebound analysis results. Probably the output of a call to `rebound_analysis()`.
+#' @param escape_latex When `TRUE` (the default), return LaTeX-compatible versions of strings.
+#' @param include_subtotals Tells whether to include rebound subtotals 
+#'                          by stage and direct/indirect.
+#'                          Default is `TRUE`.
+#' @param include_total Tells whether to include rebound total.
+#'                          Default is `TRUE`.
+#' @param as_percent When `TRUE`, rebound results are reported as percentages.
+#'                   When `FALSE`, rebound results are reported as fractions.
+#'                   Default is `TRUE`.
+#' @param rebound_terms See `ReboundTools::rebound_terms`.
+#' @param latex_rebound_terms See `ReboundTools::latex_rebound_terms`.
+#' @param case See `ReboundTools::eeu_base_params`.
+#' @param subtotals The rebound terms that represent subtotals. 
+#'                  Default is `c(ReboundTools::rebound_terms$Re_empl, ReboundTools::rebound_terms$Re_sub, ReboundTools::rebound_terms$Re_inc, ReboundTools::rebound_terms$Re_prod,  ReboundTools::rebound_terms$Re_d, ReboundTools::rebound_terms$Re_i)`.
+#' @param total The rebound term that represents total rebound. Default is `ReboundTools::rebound_terms$Re_tot`.
+#' @param term_name The title of the rebound term column. Default is "Rebound term".
+#' @param Re_val_colname The title of the rebound value column. Default is "Value \[--\]".
+#' @param perc_Re_val_colname The title of the rebound value column when percentages are requested. Default is "Value \[%\]".
+#' @param latex_term_name The LaTeX term name column. Default is "LaTeX rebound term".
+#' @param latex_Re_val_colname The LaTeX rebound value column name. Default is "Value \[--\]".
+#' @param latex_perc_Re_val_colname The LaTeX rebound value column name when percentages are requested. Default is "Value \[\\%\]".
+#' @param ... Arguments passed to `xtable::xtable()`, possibly
+#'            `label`, `caption`, `digits`, etc.
+#'
+#' @return An `xtable` object suitable for printing.
+#' 
+#' @export
+#'
+#' @examples
+#' load_eeu_data() %>% 
+#'   rebound_analysis() %>% 
+#'   rebound_results_table()
+rebound_results_table <- function(.analysis_data, 
+                                  escape_latex = TRUE,
+                                  include_subtotals = TRUE,
+                                  include_total = TRUE,
+                                  as_percent = TRUE,
+                                  rebound_terms = ReboundTools::rebound_terms,
+                                  latex_rebound_terms = ReboundTools::latex_rebound_terms,
+                                  case = ReboundTools::eeu_base_params$case, 
+                                  subtotals = c(ReboundTools::rebound_terms$Re_empl,
+                                                ReboundTools::rebound_terms$Re_sub,
+                                                ReboundTools::rebound_terms$Re_inc,
+                                                ReboundTools::rebound_terms$Re_prod, 
+                                                ReboundTools::rebound_terms$Re_d,
+                                                ReboundTools::rebound_terms$Re_i),
+                                  total = ReboundTools::rebound_terms$Re_tot,
+                                  term_name = "Rebound term", 
+                                  latex_term_name = "LaTeX rebound term",
+                                  Re_val_colname = "Value [-]",
+                                  perc_Re_val_colname = "Value [%]",
+                                  latex_Re_val_colname = "Value [--]",
+                                  latex_perc_Re_val_colname = "Value [\\%]",
+                                  ...) {
+  
+  table_data <- .analysis_data %>% 
+    dplyr::select(dplyr::any_of(case), 
+                  dplyr::any_of(rebound_terms %>% unlist())) %>% 
+    tidyr::pivot_longer(cols = rebound_terms %>% unlist(),
+                        names_to = term_name, 
+                        values_to = Re_val_colname)
+  if (!include_subtotals) {
+    table_data <- table_data %>% 
+      dplyr::filter(! .data[[term_name]] %in% subtotals)
+  }
+  if (!include_total) {
+    table_data <- table_data %>% 
+      dplyr::filter(! .data[[term_name]] %in% total)
+  }
+  if (as_percent) {
+    table_data <- table_data %>% 
+      dplyr::mutate(
+        "{Re_val_colname}" := .data[[Re_val_colname]] * 100
+      )
+  }
+  
+  if (escape_latex) {
+    # Change strings in .term_name column to be the LaTeX version.
+    latex_names <- tibble::tibble("{term_name}" := names(latex_rebound_terms),
+                                  "{latex_term_name}" := latex_rebound_terms %>% unlist())
+    table_data <- dplyr::left_join(table_data, latex_names, by = term_name) %>%
+      dplyr::mutate(
+        "{term_name}" := NULL
+      ) %>%
+      dplyr::rename(
+        "{term_name}" := .data[[latex_term_name]]
+      ) %>%
+      dplyr::relocate(.data[[term_name]], .before = .data[[Re_val_colname]])
+  }
+  
+  # Adjust the value column name if needed.
+  if (escape_latex & as_percent) {
+    table_data <- table_data %>% 
+      dplyr::rename(
+        "{latex_perc_Re_val_colname}" := .data[[Re_val_colname]]
+      )
+  } else if (as_percent) {
+    table_data <- table_data %>% 
+      dplyr::rename(
+        "{perc_Re_val_colname}" := .data[[Re_val_colname]]
+      )
+    
+  }
+  
+  # Create the xtable and return.
+  table_data %>% 
+    xtable::xtable(...)
+}